fix cache_path option to allow multiple input for intended shell behavior
Current argparse option using `type=list` does not work well with shell
as it takes in the shell string argument and split into a list of
characters, which is not the intended behavior.

reference: https://stackoverflow.com/a/15753721/498031

Quoting from the above stackoverflow link's parse output:
```
$ python arg.py --list-type "1234 2345 3456 4567"
['1', '2', '3', '4', ' ', '2', '3', '4', '5', ' ', '3', '4', '5', '6',
' ', '4', '5', '6', '7']

$ python arg.py --list-type-nargs 1234 2345 3456 4567
[['1', '2', '3', '4'], ['2', '3', '4', '5'], ['3', '4', '5', '6'], ['4',
'5', '6', '7']]

$ python arg.py --nargs 1234 2345 3456 4567
['1234', '2345', '3456', '4567']
```

Using `nargs="+"` gives us the intended behavior in the last example
above.